+++ surface/model_openapiv3.go
@@ -15,24 +15,29 @@
 package surface_v1
 
 import (
-	"errors"
-	"fmt"
-	"log"
-
-	"strings"
-
 	openapiv3 "github.com/googleapis/gnostic/OpenAPIv3"
+	"log"
+	nethttp "net/http"
+	"strconv"
+	"strings"
 )
 
-var knownTypes = map[string]bool{"string": true, "integer": true, "number": true, "boolean": true, "array": true, "object": true}
+type OpenAPI3Builder struct {
+	model *Model
+}
+
+type FieldInfo struct {
+	fieldKind   FieldKind
+	fieldType   string
+	fieldFormat string
+	// For parameters
+	fieldPosition Position
+	fieldName     string
+}
 
 // NewModelFromOpenAPIv3 builds a model of an API service for use in code generation.
 func NewModelFromOpenAPI3(document *openapiv3.Document) (*Model, error) {
 	return newOpenAPI3Builder().buildModel(document)
-}
-
-type OpenAPI3Builder struct {
-	model *Model
 }
 
 func newOpenAPI3Builder() *OpenAPI3Builder {
@@ -40,130 +45,65 @@ func newOpenAPI3Builder() *OpenAPI3Builder {
 }
 
 func (b *OpenAPI3Builder) buildModel(document *openapiv3.Document) (*Model, error) {
-	// Set model properties from passed-in document.
-	b.model.Name = document.Info.Title
 	b.model.Types = make([]*Type, 0)
 	b.model.Methods = make([]*Method, 0)
-	err := b.build(document)
-	if err != nil {
-		return nil, err
-	}
+	// Set model properties from passed-in document.
+	b.model.Name = document.Info.Title
+	b.buildFromDocument(document)
 	return b.model, nil
 }
 
 // build builds an API service description, preprocessing its types and methods for code generation.
-func (b *OpenAPI3Builder) build(document *openapiv3.Document) (err error) {
-	err = b.buildTypesFromComponents(document.Components)
-	if err != nil {
-		return err
-	}
-
-	// Collect service method descriptions from each PathItem.
-	if document.Paths != nil {
-		for _, pair := range document.Paths.Path {
-			b.buildMethodFromPathItem(pair.Name, pair.Value)
-		}
-	}
-	return err
+func (b *OpenAPI3Builder) buildFromDocument(document *openapiv3.Document) {
+	b.buildFromComponents(document.Components)
+	b.buildFromPaths(document.Paths)
 }
 
-// buildTypesFromComponents builds multiple service type description from the "Components" section
-// in the OpenAPI specification.
-func (b *OpenAPI3Builder) buildTypesFromComponents(components *openapiv3.Components) (err error) {
+func (b *OpenAPI3Builder) buildFromComponents(components *openapiv3.Components) {
 	if components == nil {
-		return nil
+		return
 	}
 
-	// Collect service type descriptions from Components/Schemas.
-	if components.Schemas != nil {
-		for _, pair := range components.Schemas.AdditionalProperties {
-			t, err := b.buildTypeFromSchemaOrReference(pair.Name, pair.Value)
-			if err != nil {
-				return err
-			}
-			if t != nil {
+	if schemas := components.Schemas; schemas != nil {
+		for _, namedSchema := range schemas.AdditionalProperties {
+			_ = b.buildFromSchemaOrReference(namedSchema.Name, namedSchema.Value)
+		}
+	}
+
+	if parameters := components.Parameters; parameters != nil {
+		for _, namedParameter := range parameters.AdditionalProperties {
+			// Parameters in OpenAPI have a name field. See: https://swagger.io/specification/#parameterObject
+			// We use that later in the buildFromParam method to name the field we build. However, a type with that
+			// parameter was never created, so we still need to do that.
+			t := b.makeType(namedParameter.Name)
+			fInfo := b.buildFromParamOrRef(namedParameter.Value)
+			b.makeFieldAndAppendToType(fInfo, t, fInfo.fieldName)
+			if len(t.Fields) > 0 {
 				b.model.addType(t)
 			}
 		}
 	}
-	// Collect service type descriptions from Components/Parameters.
-	if components.Parameters != nil {
-		for _, pair := range components.Parameters.AdditionalProperties {
-			parameters := []*openapiv3.ParameterOrReference{pair.Value}
-			_, err := b.buildTypeFromParameters(pair.Name, parameters, nil, true)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	// Collect service type descriptions from Components/requestBodies
-	if components.RequestBodies != nil {
-		for _, pair := range components.RequestBodies.AdditionalProperties {
-			t, err := b.buildTypeFromRequestBody(pair.Name, pair.Value, nil)
 
-			if err != nil {
-				return err
-			}
-			if t != nil {
-				b.model.addType(t)
-			}
-		}
-	}
-	// Collect service type descriptions from Components/responses
-	if components.Responses != nil {
-		for _, pair := range components.Responses.AdditionalProperties {
-			namedResponseOrReference := []*openapiv3.NamedResponseOrReference{pair}
-			responses := &openapiv3.Responses{ResponseOrReference: namedResponseOrReference}
-			b.buildTypeFromResponses(pair.Name, responses, true)
+	if responses := components.Responses; responses != nil {
+		for _, namedResponses := range responses.AdditionalProperties {
+			b.buildFromResponseOrRef(namedResponses.Name, namedResponses.Value)
 		}
 	}
 
-	return err
-}
-
-// buildTypeFromSchemaOrReference builds a service type description from a schema in the API description.
-func (b *OpenAPI3Builder) buildTypeFromSchemaOrReference(
-	name string,
-	schemaOrReference *openapiv3.SchemaOrReference) (t *Type, err error) {
-	if schema := schemaOrReference.GetSchema(); schema != nil {
-		t = &Type{}
-		t.Name = name
-		t.Description = "implements the service definition of " + name
-		t.Fields = make([]*Field, 0)
-		if schema.Properties != nil {
-			if len(schema.Properties.AdditionalProperties) > 0 {
-				// If the schema has properties, generate a struct.
-				t.Kind = TypeKind_STRUCT
-			}
-			for _, pair := range schema.Properties.AdditionalProperties {
-				if schema := pair.Value; schema != nil {
-					f := &Field{
-						Name:      pair.Name,
-						Serialize: true,
-					}
-					f.Kind, f.Type, f.Format = b.typeForSchemaOrReference(schema)
-					t.addField(f)
-				}
-			}
+	if requestBodies := components.RequestBodies; requestBodies != nil {
+		for _, namedRequestBody := range requestBodies.AdditionalProperties {
+			b.buildFromRequestBodyOrRef(namedRequestBody.Name, namedRequestBody.Value)
 		}
-		if len(t.Fields) == 0 {
-			if schema.AdditionalProperties != nil {
-				// If the schema has no fixed properties and additional properties of a specified type,
-				// generate a map pointing to objects of that type.
-				t.Kind = TypeKind_OBJECT
-				t.ContentType = typeForRef(schema.AdditionalProperties.GetSchemaOrReference().GetReference().GetXRef())
-			}
-		}
-		return t, err
-	} else {
-		return nil, errors.New("unable to determine service type for referenced schema " + name)
 	}
 }
 
-// buildMethodFromOperation builds a service method description
-func (b *OpenAPI3Builder) buildMethodFromPathItem(
-	path string,
-	pathItem *openapiv3.PathItem) (err error) {
+func (b *OpenAPI3Builder) buildFromPaths(paths *openapiv3.Paths) {
+	for _, path := range paths.Path {
+		b.buildFromNamedPath(path.Name, path.Value)
+	}
+}
+
+func (b *OpenAPI3Builder) buildFromNamedPath(name string, pathItem *openapiv3.PathItem) {
 	for _, method := range []string{"GET", "PUT", "POST", "DELETE", "OPTIONS", "HEAD", "PATCH", "TRACE"} {
 		var op *openapiv3.Operation
 		switch method {
@@ -187,250 +127,261 @@ func (b *OpenAPI3Builder) buildMethodFromPathItem(
 		if op != nil {
 			m := &Method{
 				Operation:   op.OperationId,
-				Path:        path,
+				Path:        name,
 				Method:      method,
 				Name:        sanitizeOperationName(op.OperationId),
 				Description: op.Description,
 			}
 			if m.Name == "" {
-				m.Name = generateOperationName(method, path)
+				m.Name = generateOperationName(method, name)
 			}
-			m.ParametersTypeName, err = b.buildTypeFromParameters(m.Name, op.Parameters, op.RequestBody, false)
-			m.ResponsesTypeName, err = b.buildTypeFromResponses(m.Name, op.Responses, false)
+			m.ParametersTypeName, m.ResponsesTypeName = b.buildFromNamedOperation(m.Name, op)
 			b.model.addMethod(m)
 		}
 	}
-	return err
 }
 
-// buildTypeFromParameters builds a service type description from the parameters of an API method
-func (b *OpenAPI3Builder) buildTypeFromParameters(
-	name string,
-	parameters []*openapiv3.ParameterOrReference,
-	requestBody *openapiv3.RequestBodyOrReference,
-	fromComponent bool) (typeName string, err error) {
+func (b *OpenAPI3Builder) buildFromNamedOperation(name string, operation *openapiv3.Operation) (parametersTypeName string, responseTypeName string) {
+	// At first, we build the operations input parameters. This includes parameters (like PATH or QUERY parameters) and a request body
+	operationParameters := b.makeType(name + "Parameters")
+	operationParameters.Description = operationParameters.Name + " holds parameters to " + name
+	for _, paramOrRef := range operation.Parameters {
+		fieldInfo := b.buildFromParamOrRef(paramOrRef)
+		// For parameters we get the name from the parameter itself. That is why we pass in fieldInfo.fieldName
+		b.makeFieldAndAppendToType(fieldInfo, operationParameters, fieldInfo.fieldName)
+	}
 
-	pName := name + "Parameters"
+	if operation.RequestBody != nil {
+		fInfo := b.buildFromRequestBodyOrRef(operation.OperationId+"RequestBody", operation.RequestBody)
+		b.makeFieldAndAppendToType(fInfo, operationParameters, "request_body")
+	}
+
+	if len(operationParameters.Fields) > 0 {
+		b.model.addType(operationParameters)
+		parametersTypeName = operationParameters.Name
+	}
+
+	// Secondly, we build the response values for the method.
+	if responses := operation.Responses; responses != nil {
+		operationResponses := b.makeType(name + "Responses")
+		operationResponses.Description = operationResponses.Name + " holds responses of " + name
+		for _, namedResponse := range responses.ResponseOrReference {
+			fieldInfo := b.buildFromResponseOrRef(operation.OperationId+convertStatusCodes(namedResponse.Name), namedResponse.Value)
+			b.makeFieldAndAppendToType(fieldInfo, operationResponses, namedResponse.Name)
+		}
+		if responses.Default != nil {
+			fieldInfo := b.buildFromResponseOrRef(operation.OperationId+"Default", responses.Default)
+			b.makeFieldAndAppendToType(fieldInfo, operationResponses, "default")
+		}
+		if len(operationResponses.Fields) > 0 {
+			b.model.addType(operationResponses)
+			responseTypeName = operationResponses.Name
+		}
+	}
+	return parametersTypeName, responseTypeName
+}
+
+func (b *OpenAPI3Builder) buildFromParamOrRef(paramOrRef *openapiv3.ParameterOrReference) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	if param := paramOrRef.GetParameter(); param != nil {
+		fInfo = b.buildFromParam(param)
+		return fInfo
+	} else if ref := paramOrRef.GetReference(); ref != nil {
+		t := findTypeForRef(b.model.Types, typeForRef(ref.XRef))
+		if t != nil {
+			fInfo.fieldKind, fInfo.fieldType, fInfo.fieldName, fInfo.fieldPosition = FieldKind_REFERENCE, typeForRef(ref.XRef), t.Name, t.Fields[0].Position
+			return fInfo
+		}
+		// TODO: This might happen for symbolic references
+		log.Printf("Not able to find parameter information for: %v", ref)
+		fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, typeForRef(ref.XRef)
+		return fInfo // Lets return fInfo for now otherwise we may get null pointer exception
+	}
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromParam(parameter *openapiv3.Parameter) (fInfo *FieldInfo) {
+	if schemaOrRef := parameter.Schema; schemaOrRef != nil {
+		fInfo = b.buildFromSchemaOrReference(parameter.Name, schemaOrRef)
+		fInfo.fieldName = parameter.Name
+		switch parameter.In {
+		case "body":
+			fInfo.fieldPosition = Position_BODY
+		case "header":
+			fInfo.fieldPosition = Position_HEADER
+		case "formdata":
+			fInfo.fieldPosition = Position_FORMDATA
+		case "query":
+			fInfo.fieldPosition = Position_QUERY
+		case "path":
+			fInfo.fieldPosition = Position_PATH
+		}
+		return fInfo
+	}
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromRequestBodyOrRef(name string, reqBodyOrRef *openapiv3.RequestBodyOrReference) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	if requestBody := reqBodyOrRef.GetRequestBody(); requestBody != nil {
+		fInfo = b.buildFromRequestBody(name, requestBody)
+		return fInfo
+	} else if ref := reqBodyOrRef.GetReference(); ref != nil {
+		fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, typeForRef(ref.XRef)
+		return fInfo
+	}
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromRequestBody(name string, reqBody *openapiv3.RequestBody) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	if reqBody.Content != nil {
+		schemaType := b.makeType(name)
+		for _, namedMediaType := range reqBody.Content.AdditionalProperties {
+			fieldInfo := b.buildFromNamedMediaType(namedMediaType.Name, namedMediaType.Value)
+			b.makeFieldAndAppendToType(fieldInfo, schemaType, namedMediaType.Name)
+		}
+		b.model.addType(schemaType)
+		fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, schemaType.Name
+		return fInfo
+	}
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromResponseOrRef(name string, responseOrRef *openapiv3.ResponseOrReference) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	if response := responseOrRef.GetResponse(); response != nil {
+		fInfo = b.buildFromResponse(name, response)
+		return fInfo
+	} else if ref := responseOrRef.GetReference(); ref != nil {
+		fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, typeForRef(ref.XRef)
+		return fInfo
+	}
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromResponse(name string, response *openapiv3.Response) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	if response.Content != nil && response.Content.AdditionalProperties != nil {
+		schemaType := b.makeType(name)
+		for _, namedMediaType := range response.Content.AdditionalProperties {
+			fieldInfo := b.buildFromNamedMediaType(name+namedMediaType.Name, namedMediaType.Value)
+			b.makeFieldAndAppendToType(fieldInfo, schemaType, namedMediaType.Name)
+		}
+		b.model.addType(schemaType)
+		fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, schemaType.Name
+		return fInfo
+	}
+	log.Printf("Response has no content: %v", name)
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromNamedMediaType(name string, mediaType *openapiv3.MediaType) (fInfo *FieldInfo) {
+	if schemaOrRef := mediaType.Schema; schemaOrRef != nil {
+		fInfo = b.buildFromSchemaOrReference(name, schemaOrRef)
+	}
+	return fInfo
+}
+
+func (b *OpenAPI3Builder) buildFromSchemaOrReference(name string, schemaOrReference *openapiv3.SchemaOrReference) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	if schema := schemaOrReference.GetSchema(); schema != nil {
+		fInfo = b.buildFromSchema(name, schema)
+		return fInfo
+	} else if ref := schemaOrReference.GetReference(); ref != nil {
+		fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, typeForRef(ref.XRef)
+		return fInfo
+	}
+	return nil
+}
+
+func (b *OpenAPI3Builder) buildFromSchema(name string, schema *openapiv3.Schema) (fInfo *FieldInfo) {
+	fInfo = &FieldInfo{}
+	// Data types according to: https://swagger.io/docs/specification/data-models/data-types/
+	switch schema.Type {
+	case "":
+		fallthrough
+	case "object":
+		schemaType := b.makeType(name)
+		if schema.Properties != nil && schema.Properties.AdditionalProperties != nil {
+			for _, namedSchema := range schema.Properties.AdditionalProperties {
+				fieldInfo := b.buildFromSchemaOrReference(namedSchema.Name, namedSchema.Value)
+				b.makeFieldAndAppendToType(fieldInfo, schemaType, namedSchema.Name)
+			}
+		}
+		if schemaOrRef := schema.AdditionalProperties.GetSchemaOrReference(); schemaOrRef != nil {
+			// AdditionalProperties are represented as map
+			fieldInfo := b.buildFromSchemaOrReference(name+"AdditionalProperties", schemaOrRef)
+			mapValueType := b.makeMapValueType(*fieldInfo)
+			fieldInfo.fieldKind, fieldInfo.fieldType, fieldInfo.fieldFormat = FieldKind_MAP, "map[string]"+mapValueType, ""
+			b.makeFieldAndAppendToType(fieldInfo, schemaType, "additional_properties")
+		}
+		if len(schemaType.Fields) > 0 {
+			b.model.addType(schemaType)
+			fInfo.fieldKind, fInfo.fieldType = FieldKind_REFERENCE, schemaType.Name
+			return fInfo
+		}
+	case "array":
+		for _, schemaOrRef := range schema.Items.SchemaOrReference {
+			arrayFieldInfo := b.buildFromSchemaOrReference("", schemaOrRef)
+			fInfo.fieldKind, fInfo.fieldType, fInfo.fieldFormat = FieldKind_ARRAY, arrayFieldInfo.fieldType, arrayFieldInfo.fieldFormat
+		}
+		return fInfo
+	default:
+		// We go a scalar value
+		fInfo.fieldKind, fInfo.fieldType, fInfo.fieldFormat = FieldKind_SCALAR, schema.Type, schema.Format
+		return fInfo
+	}
+	log.Printf("Unimplemented: could not find field info for schema: %v", schema)
+	return nil
+}
+
+func (b *OpenAPI3Builder) makeType(name string) *Type {
 	t := &Type{
-		Name:        pName,
-		Kind:        TypeKind_STRUCT,
-		Fields:      make([]*Field, 0),
-		Description: pName + " holds parameters to " + name,
+		Name:   name,
+		Kind:   TypeKind_STRUCT,
+		Fields: make([]*Field, 0),
 	}
-
-	if fromComponent {
-		t.Name = name
-		t.Description = t.Name + " is a parameter"
-	}
-	for _, parametersItem := range parameters {
-		f := Field{
-			Type: fmt.Sprintf("%+v", parametersItem),
-		}
-		parameter := parametersItem.GetParameter()
-		if parameter != nil {
-			switch parameter.In {
-			case "body":
-				f.Position = Position_BODY
-			case "header":
-				f.Position = Position_HEADER
-			case "formdata":
-				f.Position = Position_FORMDATA
-			case "query":
-				f.Position = Position_QUERY
-			case "path":
-				f.Position = Position_PATH
-			}
-			f.Name = parameter.Name
-			if parameter.GetSchema() != nil && parameter.GetSchema() != nil {
-				f.Kind, f.Type, f.Format = b.typeForSchemaOrReference(parameter.GetSchema())
-			}
-			f.Serialize = true
-			t.addField(&f)
-		} else if parameterRef := parametersItem.GetReference(); parameterRef != nil {
-			f.Type = typeForRef(parameterRef.GetXRef())
-			f.Name = strings.ToLower(f.Type)
-			f.Kind = FieldKind_REFERENCE
-			f.Position = b.positionForType(f.Type)
-
-			t.addField(&f)
-		}
-	}
-
-	_, err = b.buildTypeFromRequestBody(name, requestBody, t)
-
-	if len(t.Fields) > 0 {
-		b.model.addType(t)
-		return t.Name, err
-	}
-	return "", err
+	return t
 }
 
-// buildTypeFromRequestBody builds a service type description from the request bodies of an OpenAPI
-// description. If tIn is not given, a new type is created. Otherwise tIn is used.
-func (b *OpenAPI3Builder) buildTypeFromRequestBody(name string, requestBody *openapiv3.RequestBodyOrReference, tIn *Type) (tOut *Type, err error) {
-	tOut = &Type{
-		Name: name,
+func (b *OpenAPI3Builder) makeFieldAndAppendToType(info *FieldInfo, schemaType *Type, fieldName string) {
+	if info != nil {
+		f := &Field{Name: fieldName}
+		f.Type, f.Kind, f.Format, f.Position = info.fieldType, info.fieldKind, info.fieldFormat, info.fieldPosition
+		schemaType.Fields = append(schemaType.Fields, f)
 	}
-
-	if tIn != nil {
-		tOut = tIn
-	}
-
-	if requestBody != nil {
-		content := requestBody.GetRequestBody().GetContent()
-		f := &Field{
-			Position:  Position_BODY,
-			Serialize: true,
-		}
-
-		if content != nil {
-			for _, pair := range content.GetAdditionalProperties() {
-				if pair.Name != "application/json" {
-					log.Printf("unimplemented: %q requestBody(%s)", name, pair.Name)
-					continue
-				}
-				f.Kind, f.Type, f.Format = b.typeForSchemaOrReference(pair.GetValue().GetSchema())
-				f.Name = strings.ToLower(f.Type) // use the schema name as the parameter name, since none is directly specified
-				tOut.addField(f)
-			}
-		} else if reference := requestBody.GetReference(); reference != nil {
-			schemaOrReference := openapiv3.SchemaOrReference{&openapiv3.SchemaOrReference_Reference{Reference: reference}}
-			f.Kind, f.Type, f.Format = b.typeForSchemaOrReference(&schemaOrReference)
-			f.Name = strings.ToLower(f.Type) // use the schema name as the parameter name, since none is directly specified
-			tOut.addField(f)
-		}
-	}
-
-	return tOut, err
 }
 
-// buildTypeFromResponses builds a service type description from the responses of an API method
-func (b *OpenAPI3Builder) buildTypeFromResponses(
-	name string,
-	responses *openapiv3.Responses,
-	fromComponent bool) (typeName string, err error) {
-
-	rName := name + "Responses"
-	t := &Type{
-		Name:        name + "Responses",
-		Kind:        TypeKind_STRUCT,
-		Fields:      make([]*Field, 0),
-		Description: rName + " holds responses of " + name,
+func (b *OpenAPI3Builder) makeMapValueType(fInfo FieldInfo) (mapValueType string) {
+	if fInfo.fieldKind == FieldKind_ARRAY {
+		mapValueType = "[]"
 	}
-
-	if fromComponent {
-		t.Name = name
-		t.Description = t.Name + " is a response"
+	if fInfo.fieldFormat != "" {
+		fInfo.fieldType = fInfo.fieldFormat
 	}
-
-	addResponse := func(name string, value *openapiv3.ResponseOrReference) {
-		f := Field{
-			Name:      name,
-			Serialize: false,
-		}
-		response := value.GetResponse()
-		if response != nil && response.GetContent() != nil {
-			for _, pair2 := range response.GetContent().GetAdditionalProperties() {
-				f.Kind, f.Type, f.Format = b.typeForSchemaOrReference(pair2.GetValue().GetSchema())
-
-				if !t.HasFieldWithName(f.Name) {
-					t.addField(&f)
-				}
-			}
-		} else if responseRef := value.GetReference(); responseRef != nil {
-			schemaOrReference := openapiv3.SchemaOrReference{&openapiv3.SchemaOrReference_Reference{Reference: responseRef}}
-			f.Kind, f.Type, f.Format = b.typeForSchemaOrReference(&schemaOrReference)
-			t.addField(&f)
-		}
-	}
-
-	for _, pair := range responses.ResponseOrReference {
-		addResponse(pair.Name, pair.Value)
-	}
-	if responses.Default != nil {
-		addResponse("default", responses.Default)
-	}
-
-	if len(t.Fields) > 0 {
-		b.model.addType(t)
-		return t.Name, err
-	}
-	return "", err
+	mapValueType += fInfo.fieldType
+	return mapValueType
 }
 
-// typeForSchemaOrReference determines the language-specific type of a schema or reference
-func (b *OpenAPI3Builder) typeForSchemaOrReference(value *openapiv3.SchemaOrReference) (kind FieldKind, typeName, format string) {
-	if value.GetSchema() != nil {
-		return b.typeForSchema(value.GetSchema())
+func convertStatusCodes(c string) (statusText string) {
+	code, err := strconv.Atoi(c)
+	if err == nil {
+		statusText = nethttp.StatusText(code)
+		if statusText == "" {
+			log.Println("Status code " + c + "is not known to net.http.StatusText. This might cause unpredictable behavior.")
+			statusText = "unknownStatusCode"
+		}
+		statusText = strings.Replace(statusText, " ", "_", -1)
 	}
-	if value.GetReference() != nil {
-		return FieldKind_REFERENCE, typeForRef(value.GetReference().XRef), ""
-	}
-	return FieldKind_SCALAR, "todo", ""
+	return statusText
 }
 
-// typeForSchema determines the language-specific type of a schema
-func (b *OpenAPI3Builder) typeForSchema(schema *openapiv3.Schema) (kind FieldKind, typeName, format string) {
-	if schema.Type != "" {
-		format := schema.Format
-		switch schema.Type {
-		case "string":
-			return FieldKind_SCALAR, "string", format
-		case "integer":
-			return FieldKind_SCALAR, "integer", format
-		case "number":
-			return FieldKind_SCALAR, "number", format
-		case "boolean":
-			return FieldKind_SCALAR, "boolean", format
-		case "array":
-			if schema.Items != nil {
-				// we have an array.., but of what?
-				items := schema.Items
-				if items != nil {
-					a := items.GetSchemaOrReference()
-					if a[0].GetReference().GetXRef() != "" {
-						return FieldKind_ARRAY, typeForRef(a[0].GetReference().GetXRef()), format
-					} else if knownTypes[a[0].GetSchema().Type] {
-						// The items of the array is one of the known types.
-						return FieldKind_ARRAY, a[0].GetSchema().Type, a[0].GetSchema().Format
-					}
-				}
-			}
-		case "object":
-			if schema.AdditionalProperties == nil {
-				return FieldKind_MAP, "object", format
-			}
-		case "any":
-			return FieldKind_ANY, "any", format
-		default:
-
+func findTypeForRef(types []*Type, typeName string) *Type {
+	for _, t := range types {
+		if typeName == t.Name {
+			return t
 		}
 	}
-	if schema.AdditionalProperties != nil {
-		schemaOrReference := schema.AdditionalProperties.GetSchemaOrReference()
-		k, t, f := b.typeForSchemaOrReference(schemaOrReference)
-		mapValueType := ""
-		if k == FieldKind_ARRAY {
-			mapValueType = "[]"
-		}
-		if f != "" {
-			t = f
-		}
-		mapValueType += t
-		return FieldKind_MAP, "map[string]" + mapValueType, ""
-	}
-	// this function is incomplete... use generic interface{} for now
-	log.Printf("unimplemented: %v", schema)
-	return FieldKind_SCALAR, "object", ""
-}
-
-// Searches all types that have been created so far for the type with 'typeName' and
-// returns the position of the first field. Returns Position_BODY if there is no such type.
-func (b *OpenAPI3Builder) positionForType(typeName string) Position {
-	for _, t := range b.model.Types {
-		if t.Name == typeName {
-			return t.Fields[0].Position
-		}
-	}
-	return Position_BODY
+	return nil
 }
